migate編集

+++ db/migrate/20221112125118_create_likes.rb
@@ -1,9 +1,9 @@
 class CreateLikes < ActiveRecord::Migration[6.1]
   def change
     create_table :likes do |t|
-      
-     t.integer :user_id
-     t.integer :post_id
+
+     t.bigint :user_id
+     t.bigint :post_id
 
       t.timestamps
     end
